requirements.txt
boto3
dash
dash-bootstrap-components
elasticsearch
Flask
gunicorn
numpy
pandas
plotly
python-dotenv
